+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - it function support in IML tests
 - Integromat Help command
 - Publish / Hide App or Module
+- IML tests with users' functions
+- Keybind and command for running IML test
 
 ### Fixed